+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add optimizer case for `nrt`
 * Add internal command `test-recursion`
 * Add internal command `update-formula-tests` to automatically update formula test expectations
+* Add program change log in server mode to track program updates (added, updated, removed) with reasons in `$LODA_HOME/debug/programs_changelog.txt`
 
 ## v25.12.1